requirements.txt
fastapi
uvicorn
ultralytics
pillow
torch
easyocr
opencv-python-headless
